requirements.txt
numpy==1.23.5
wilds==1.2.2
imageio==2.9.0
gdown==3.13.0
tqdm==4.62.2
backpack==0.1
parameterized==0.8.1
Pillow==8.3.2
torchmetrics==1.3.2
matplotlib==3.9.0
seaborn==0.13.2
transformers==4.41.2
chardet==5.2.0
accelerate==1.2.1
openai==1.57.0
nest-asyncio==1.6.0
h5py==3.12.1
opencv-python
grad-cam